# internal/swagger/swagger.yaml
definitions:
...
  github_com_tokenlive_tokenlive-admin_pkg_errors.Error:
    properties:
      code:
        type: integer
      detail:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  schema.Estimator:
    properties:
      ratio:
        description: Character to token ratio
        type: number
      type:
        description: length_ratio, tiktoken
        type: string
    type: object
...
  schema.OpType:
    enum:
    - EQUAL
    - NOT_EQUAL
    - IN
    - NOT_IN
    - REGULAR
    - PREFIX
    type: string
    x-enum-varnames:
    - OpType_EQUAL
    - OpType_NOT_EQUAL
    - OpType_IN
    - OpType_NOT_IN
    - OpType_REGULAR
    - OpType_PREFIX
...
  schema.PolicyCopyToModelForm:
    properties:
      model_id:
        maxLength: 20
        type: string
      name:
        maxLength: 128
        type: string
      priority:
        minimum: 0
        type: integer
      scope_code:
        maxLength: 128
        type: string
      scope_type:
        maxLength: 32
        type: string
    required:
    - model_id
    type: object
...
  schema.PolicyLimit:
    properties:
      conditions:
        type: string
      created_at:
        type: string
      creator:
        type: string
      description:
        type: string
      enabled:
        type: integer
      estimator:
        type: string
      id:
        type: string
      limit_by:
        type: string
      max_wait_ms:
        type: integer
      model_id:
        type: string
      modifier:
        type: string
      name:
        type: string
      priority:
        type: integer
      relation_type:
        type: string
      scope_code:
        type: string
      scope_type:
        type: string
      sliding_windows:
        type: string
      type:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  schema.PolicyLimitForm:
    properties:
      conditions:
        description: Match conditions (JSON)
        items:
          $ref: '#/definitions/schema.TagCondition'
        type: array
      created_at:
        description: Create timestamp
        type: string
      creator:
        description: Creator
        type: string
      description:
        description: Details
        type: string
      enabled:
        description: Enabled
        type: integer
      estimator:
        allOf:
        - $ref: '#/definitions/schema.Estimator'
        description: Estimator config (JSON)
      id:
        type: string
      limit_by:
        description: Limit key dimensions (JSON)
        items:
          type: string
        type: array
      max_wait_ms:
        description: Max queue wait time (ms)
        type: integer
      model_id:
        description: Owner model ID; empty means template
        maxLength: 20
        type: string
      modifier:
        description: Modifier
        type: string
      name:
        description: Policy name
        maxLength: 128
        type: string
      priority:
        description: Priority
        minimum: 0
        type: integer
      relation_type:
        description: 'Relation type: AND / OR'
        maxLength: 16
        type: string
      scope_code:
        description: Scope code
        maxLength: 128
        type: string
      scope_type:
        description: Scope type
        maxLength: 32
        type: string
      sliding_windows:
        description: Sliding windows (JSON)
        items:
          $ref: '#/definitions/schema.SlidingWindow'
        type: array
      type:
        description: 'Limit dimension: request / token / cost'
        maxLength: 64
        type: string
      updated_at:
        description: Update timestamp
        type: string
      version:
        description: Version
        type: integer
    required:
    - name
    - relation_type
    - type
    type: object
  schema.PolicyLoadbalance:
    properties:
      created_at:
        type: string
      creator:
        type: string
      description:
        type: string
      enabled:
        type: integer
      id:
        type: string
      model_id:
        type: string
      modifier:
        type: string
      name:
        type: string
      params:
        items:
          type: integer
        type: array
      priority:
        type: integer
      scope_code:
        type: string
      scope_type:
        type: string
      type:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  schema.PolicyLoadbalanceForm:
    properties:
      created_at:
        description: Create timestamp
        type: string
      creator:
        description: Creator
        type: string
      description:
        description: Details
        type: string
      enabled:
        description: Enabled
        type: integer
      id:
        type: string
      model_id:
        description: Owner model ID; empty means template
        maxLength: 20
        type: string
      modifier:
        description: Modifier
        type: string
      name:
        description: Policy name
        maxLength: 128
        type: string
      params:
        description: Extra params (JSON)
        items:
          type: integer
        type: array
      priority:
        description: Priority
        minimum: 0
        type: integer
      scope_code:
        description: Scope code
        maxLength: 128
        type: string
      scope_type:
        description: Scope type
        maxLength: 32
        type: string
      type:
        description: Loadbalance policy type
        maxLength: 64
        type: string
      updated_at:
        description: Update timestamp
        type: string
      version:
        description: Version
        type: integer
    required:
    - name
    - type
    type: object
...
  schema.SlidingWindow:
    properties:
      burst_ratio:
        type: number
      threshold:
        type: integer
      time_window_in_ms:
        type: integer
    type: object
...
  schema.TagCondition:
    properties:
      key:
        type: string
      op_type:
        $ref: '#/definitions/schema.OpType'
      type:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_tokenlive_tokenlive-admin_pkg_errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
info:
  contact: {}
  description: An admin control center for tokelvie.
  title: tokenlive-admin
  version: v1.0.0
paths:
...
  /api/v1/policy/policy-limits:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.PolicyLimit'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query policy limit list
      tags:
      - PolicyLimitAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyLimitForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyLimit'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create policy limit record
      tags:
      - PolicyLimitAPI
  /api/v1/policy/policy-limits/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete policy limit record by ID
      tags:
      - PolicyLimitAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyLimitForm'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get policy limit record by ID
      tags:
      - PolicyLimitAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyLimitForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update policy limit record by ID
      tags:
      - PolicyLimitAPI
  /api/v1/policy/policy-limits/{id}/copy-to-model:
    post:
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyCopyToModelForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyLimit'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Copy policy limit template to model
      tags:
      - PolicyLimitAPI
  /api/v1/policy/policy-loadbalances:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.PolicyLoadbalance'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query policy loadbalance list
      tags:
      - PolicyLoadbalanceAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyLoadbalanceForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyLoadbalance'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create policy loadbalance record
      tags:
      - PolicyLoadbalanceAPI
  /api/v1/policy/policy-loadbalances/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete policy loadbalance record by ID
      tags:
      - PolicyLoadbalanceAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyLoadbalanceForm'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get policy loadbalance record by ID
      tags:
      - PolicyLoadbalanceAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyLoadbalanceForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update policy loadbalance record by ID
      tags:
      - PolicyLoadbalanceAPI
  /api/v1/policy/policy-loadbalances/{id}/copy-to-model:
    post:
      parameters:
      - description: template id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.PolicyCopyToModelForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.PolicyLoadbalance'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Copy policy loadbalance template to model
      tags:
      - PolicyLoadbalanceAPI
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"